Refactor: 테스트 파일 디렉토리 변경
